requirements.txt
numpy
pandas
scikit-learn
nltk
spacy
PyPDF2
python-docx
django
gunicorn
whitenoise
dj-database-url
psycopg2-binary
openpyxl
reportlab
python-dotenv
